requirements.txt
asgiref==3.7.2
cffi==1.16.0
cryptography==42.0.1
Django==5.0.1
django-filter==23.5
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
pycparser==2.21
PyJWT==2.8.0
pytz==2023.3.post1
simplejwt==2.0.1
sqlparse==0.4.4
typing==3.7.4.3
typing_extensions==4.9.0
